refactor: [BE] 주문하기 - 가격비교, 총 가격 계산 함수 order.entity로 분리
